% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-SparkR.r
\name{bindIDs}
\alias{bindIDs}
\title{This function is only for Spark DataFrame}
\usage{
bindIDs(data)
}
\arguments{
\item{data}{A Spark DataFrame}
}
\value{
A DataFrame containing the result of the binding with IDs
}
\description{
Return a new DataFrame containing the binding of ID with original DataFrame.
This method only used by one-column DataFrame now.
}

