% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-null.r
\name{facet_train_layout_SparkR.null}
\alias{facet_train_layout_SparkR.null}
\title{Facet spaecification: a single panel.}
\usage{
\method{facet_train_layout_SparkR}{null}(facet, data)
}
\description{
Facet spaecification: a single panel.
}

