% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bar.r, R/stat-count.r
\docType{data}
\name{geom_bar}
\alias{GeomBar_SparkR}
\alias{StatCount_SparkR}
\alias{geom_bar}
\alias{stat_count}
\title{Bars, rectangles with bases on x-axisi}
\usage{
geom_bar(mapping = NULL, data = NULL, stat = "count",
  position = "stack", width = NULL, binwidth = NULL, ..., na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Override the default connection between geom_bar and stat_count}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{width}{Bar width.}

\item{binwidth}{geom_bar no longer has a binwidth argument}

\item{...}{other arguments passed on to layer.}

\item{na.rm}{If FALSE, removes missing values with a warning (not used in ggplot2.SparkR).}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}
}
\description{
There are two types of bar charts, determined by what is mapped to bar
height. By default, geom_bar uses stat_count which makes the
height of the bar proportion to the number of cases.
}
\details{
A bar chart maps the height of the bar to a variable, and so the base of the
bar must always be shown to produce a valid visual comparison. Naomi Robbins
has a nice
\href{http://www.b-eye-network.com/view/index.php?cid=2468}{article on this
topic}. This is why it doesn't make sense to use a log-scaled y axis with a
bar chart.
}
\examples{
\dontrun{
ggplot(faithful, aes(x = eruptions)) + geom_bar()

df <- createDataFrame(sqlContext, faithful)
ggplot(df, aes(x = eruptions)) + geom_bar()

df2 <- createDataFrame(sqlContext, diamonds)
ggplot(df2, aes(x = cut, fill = color)) + geom_bar(position = "dodge")
ggplot(df2, aes(cut)) + geom_bar() + facet_grid(. ~ clarity)
}
}
\keyword{datasets}

