% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-bin2d.r, R/geom-tile.r, R/stat-bin2d.r
\docType{data}
\name{geom_bin2d}
\alias{GeomTile_SparkR}
\alias{StatBin2d_SparkR}
\alias{geom_bin2d}
\alias{stat_bin2d}
\alias{stat_bin_2d}
\title{Add heatmap of 2d bin counts.}
\usage{
geom_bin2d(mapping = NULL, data = NULL, stat = "bin2d",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Use override the default connection between geom_bin2d and stat_bin2d}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If FALSE, remove missing values with a warning (not used in ggplot2.SparkR)}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}

\item{...}{other arguments passed on to layer.}
}
\description{
Add heatmap of 2d bin counts.

Count number of observation in rectangular bins.
}
\examples{
\dontrun{
ggplot(diamonds, aes(x, y)) + geom_bin2d()

df <- createDataFrame(sqlContext, diamonds)
ggplot(df, aes(x, y)) + geom_bin2d()
ggplot(df, aes(x, y)) + geom_bin2d(binwidth = c(0.5, 0.5))
}
}
\keyword{datasets}

