% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-boxplot.r, R/stat-boxplot.r
\docType{data}
\name{geom_boxplot}
\alias{GeomBoxplot_SparkR}
\alias{StatBoxplot_SparkR}
\alias{geom_boxplot}
\alias{stat_boxplot}
\title{Box and whiskers plot.}
\usage{
geom_boxplot(mapping = NULL, data = NULL, stat = "boxplot",
  position = "dodge", outlier.colour = NULL, outlier.shape = 19,
  outlier.size = 1.5, outlier.stroke = 0.5, notch = FALSE,
  notchwidth = 0.5, varwidth = FALSE, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Use to override the default connection between geom_boxplot and stat_boxplot}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{outlier.colour, outlier.shape, outlier.size, outlier.stroke}{Default aesthetics for outliers.}

\item{notch}{if FALSE(default) make a standard box plot.}

\item{notchwidth}{for a notched box plot, width of the notch relative to
the body (default 0.5)}

\item{varwidth}{if FALSE(default) make a standard box plot.}

\item{na.rm}{If FALSE, removes missing values with a warning (not used in ggplot2.SparkR).}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}

\item{...}{other arguments passed on to layer.}

\item{geom}{The geometric object to use display the data}
}
\description{
The lower and upper "hinges" correspond to the first and third quartiles
(the 25th and 75th percentiles). This differs slightly from the method used
by the boxplot function, and may be apparent with small samples.
}
\details{
The upper whisker extends from the hinge to the highest value that is within
1.5 * IQR of the hinge, where IQR is the inter-quartile range, or distance
between the first and third quartiles. The lower whisker extends from the
hinge to the lowest value within 1.5 * IQR of the hinge. Data beyond the
end of the whiskers are outliers and plotted as points (as specified by Tukey).

In a notched box plot, the notches extend \code{1.58 * IQR / sqrt(n)}.
This gives a roughly 95% confidence interval for comparing medians.
See McGill et al. (1978) for more details.
}
\examples{
\dontrun{
ggplot(diamonds, aes(cut, price)) + geom_boxplot()

df <- createDataFrame(sqlContext, diamonds)
ggplot(df, aes(cut, price)) + geom_bar()
ggplot(df, aes(cut, price, fill = color)) + geom_boxplot()
}
}
\keyword{datasets}

