% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-histogram.r, R/stat-bin.r
\docType{data}
\name{geom_histogram}
\alias{StatBin_SparkR}
\alias{geom_histogram}
\alias{stat_bin}
\title{Histograms.}
\usage{
geom_histogram(mapping = NULL, data = NULL, stat = "bin",
  binwidth = NULL, bins = NULL, origin = NULL, right = FALSE,
  position = "stack", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{Use override the default connection between geom_histogram, geom_freqpoly and stat_bin}

\item{binwidth}{???}

\item{bins}{Does not use in ggplot2.SparkR}

\item{origin}{Does not use in ggplot2.SparkR}

\item{right}{Does not use in ggplot2.SparkR}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{Not used in ggplot2.SparkR}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}

\item{...}{other arguments passed on to layer.}
}
\description{
Display a 1d distribution by dividing into bins and counting the number
of observations in each bin. Histograms use bars.

\code{stat_bin} is suitable only for continuous x data. If your x data is
  discrete, you probably want to use \code{\link{stat_count}}.
}
\examples{
\dontrun{
ggplot(faithful, aes(eruptions)) + geom_histogram()

df <- createDataFrame(sqlContext, faithful)
ggplot(df, aes(eruptions)) + geom_histogram()
ggplot(df, aes(eruptions)) + geom_histogram(binwidth = 5)

ggplot(df, aes(eruptions, fill = waiting)) + geom_histogram()
}
}
\keyword{datasets}

