% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{ggplot.DataFrame}
\alias{ggplot.DataFrame}
\title{Create a new ggplot plot from a Spark DataFrame}
\usage{
\method{ggplot}{DataFrame}(data, mapping = aes(), ...,
  environment = parent.frame())
}
\arguments{
\item{data}{default DataFrame for plot}

\item{mapping}{default list of aesthetic mappins (these can be colour,
size, shape, line type -- see individual geom functions for more details)}

\item{...}{ignored}

\item{environment}{in which evaluation of aesthetics should occur}
}
\description{
Create a new ggplot plot from a Spark DataFrame
}

