% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.r
\name{ggplot_build.SparkR}
\alias{ggplot_build.SparkR}
\title{Build ggplot for rendering using Spark DataFrame.}
\usage{
ggplot_build.SparkR(plot)
}
\arguments{
\item{plot}{ggplot.SparkR object}
}
\description{
This function takes the plot object, and performs all steps necessary to
produce an object that can be rendered.  This function outputs two pieces:
a list of DataFrames (one for each layer), and a panel object, which
contain all information about axis limits, breaks etc.
}
\seealso{
\code{\link{print.ggplot.SparkR}} and \code{\link{benchplot}}
 for functions that contain the complete set of steps for generating
 a ggplot2 plot.
}
\keyword{internal}

