% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.r
\name{layer_SparkR}
\alias{layer_SparkR}
\title{Create a new layer}
\usage{
layer_SparkR(geom = NULL, stat = NULL, data = NULL, mapping = NULL,
  position = NULL, params = list(), inherit.aes = TRUE, subset = NULL,
  show.legend = NA)
}
\arguments{
\item{geom}{The geometric object to use display the data}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{data}{The data to be displayed in this layer.}

\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{params}{Additional parameters to the geom and stat.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}

\item{subset}{DEPRECATED. Do not use now.}

\item{show.legend}{logical. Should this layer be included in the legends?}
}
\description{
A layer is a combination of data, stat and geom with a potential position
adjustment. Usually layers are created using \code{geom_*} or \code{stat_*}
calls but it can also be created directly using this function.
}

