% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.r
\name{limits}
\alias{limits}
\title{Generate correct scale type for specified limits}
\usage{
limits(lims, var)
}
\arguments{
\item{var}{name of variable}

\item{limts}{vector of limits}
}
\description{
Generate correct scale type for specified limits
}
\examples{
ggplot2.SparkR:::limits(c(1, 5), "x")
ggplot2.SparkR:::limits(c(5, 1), "x")
ggplot2.SparkR:::limits(c("A", "b", "c"), "x")
ggplot2.SparkR:::limits(c("A", "b", "c"), "fill")
ggplot2.SparkR:::limits(as.Date(c("2008-01-01", "2009-01-01")), "x")
}
\keyword{internal}

