% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-render.r
\name{print.ggplot.SparkR}
\alias{plot.ggplot.SparkR}
\alias{print.ggplot.SparkR}
\title{Draw plot on currernt graphics device with Spark DataFrame.}
\usage{
\method{print}{ggplot.SparkR}(x, newpage = is.null(vp), vp = NULL, ...)

\method{plot}{ggplot.SparkR}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display that has Spark DataFrame}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\description{
Draw plot on currernt graphics device with Spark DataFrame.
}
\keyword{hplot}

