% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-point.r, R/stat-sum.r
\docType{data}
\name{GeomPoint_SparkR}
\alias{GeomPoint_SparkR}
\alias{StatSum_SparkR}
\alias{stat_sum}
\title{Count two dimensional variables and present size of point}
\usage{
stat_sum(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mapping created by aes or aes_.}

\item{data}{The data to be displayed in this layer.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If FALSE, removes missing values with a warning (not used in ggplot2.SparkR).}

\item{show.legend}{logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather ehan combining with them.}

\item{...}{other arguments passed on to layer.}
}
\description{
Count two dimensional variables and present size of point
}
\keyword{datasets}

